\name{mdl}
\alias{mdl}
\docType{data}
\title{
Water data set: matrix of limits of detection 
}
\description{
Matrix of varying limits of detection for the \code{\link{Water}} data set.
}
\usage{data(mdl)}
\format{
  A \code{\link{matrix}} with 100 rows and 4 columns.
}

\details{Three limits of detection (0.75, 1 and 1.25) were considered for Potassium, four for Arsenic (1.5, 3, 4 and 5), two for Sulphate (29 and 35) and no one for Calcium.}

\examples{
data(Water)
data(mdl)
}
\keyword{datasets}
