% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perLog.R
\name{perLog}
\alias{perLog}
\title{Test of differences in group means for compositional data}
\usage{
perLog(X, groups = NULL, p = 10, alpha = 0.05, R = 1000)
}
\arguments{
\item{X}{Compositional data set (\code{\link{matrix}} or \code{\link{data.frame}} class).}

\item{groups}{Factor variable indicating the grouping structure. If NULL (default), any zero patterns in the data will be used as internal grouping factor.

Note that if a grouping factor is set by the user, then any zeros in the data must be previously dealt with, e.g. by imputation.}

\item{p}{Power parameter used in overall dissimilarity test statistic. (default = 10).}

\item{alpha}{Statistical significance level (default = 0.05).}

\item{R}{Number of permutation resamples (default = 1000).}
}
\value{
A list object of class "printres" containing:
\item{disOv}{Overall dissimilarity measure (E).}
\item{pvalOv}{Overall permutation p-value.}
}
\description{
A nonparametric permutation test based on pairwise logratios to assess the null hypothesis of equality of means/locations between subsets of compositions according to an externally or internally defined factor. If any, zero patterns are considered as default internal grouping factor.
}
\examples{
# Load the Water data set 
data(Water)
# Visualise zero patterns and select the first three for illustration
tmp <- zPatterns(Water, label = 0)
Water2 <- Water[tmp \%in\% c(1,2,3),]
# Perform the perLog test on the selected data set
perLog(Water2)
}
\references{
Aitchison, J. (1986). The Statistical Analysis of Compositional Data. Chapman and Hall, London.
}
\seealso{
\code{\link{zPatterns}}
}
