\name{Pigs}
\alias{Pigs}
\docType{data}
\title{
Pigs data set
}
\description{
Count data set consisting of scan sample behavioural observations of a group of 29 sows during a day from 7:30am to 3:30pm, and recorded every 5 minutes (97 times). Six locations were considered: straw bed (BED), half in the straw bed (HALF.BED), dunging passage (PASSAGE), half in the dunging passage (HALF.PASS), feeder (FEEDER) and half in the feeder (HALF.FEED).
}
\usage{data(Pigs)}
\format{
  A \code{\link{data.frame}} with 29 observations on the following 6 variables.
  \describe{
    \item{\code{BED}}{a numeric vector}
    \item{\code{HALF.BED}}{a numeric vector}
    \item{\code{PASSAGE}}{a numeric vector}
    \item{\code{HALF.PASS}}{a numeric vector}
    \item{\code{FEEDER}}{a numeric vector}
    \item{\code{HALF.FEED}}{a numeric vector}
  }
}
\source{
Data set kindly provided by the Animal Behaviour and Welfare group at Scotland's Rural College (SRUC), Scotland, UK.
}

\examples{
data(Pigs)
}
\keyword{datasets}
