% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcTest.R
\name{lcTest}
\alias{lcTest}
\title{Log-contrast homogeneity test}
\usage{
lcTest(X, label = 0, groups = NULL, lc = NULL, b = 1000)
}
\arguments{
\item{X}{Compositional data set (\code{\link{matrix}} or \code{\link{data.frame}} class).}

\item{label}{Unique label (\code{\link{numeric}} or \code{\link{character}}) used to denote zero or unobserved data in \code{X} (\code{label = 0}, default).}

\item{groups}{Grouping factor in fully observed zero-free data sets (\code{groups = NULL}, default).}

\item{lc}{User-defined log-contrast (see details below).}

\item{b}{Number of bootstrap resamples used (\code{b = 1000}, default).}
}
\value{
Test p-values for log-contrast variances and means.
}
\description{
This function performs tests of the homogeneity across groups of means and variances of
user-defined log-contrasts. Groups can be defined by either zero/unobserved data patterns or by a grouping
factor in fully observed zero-free data sets.
}
\details{
Non-parametric Kruskal-Wallis and Fligner-Killen test are used for testing homogeneity of
means and variances respectively of the log-contrast across groups. A permutation test of homogeneity is also
conducted based on total weighted squared relative errors (see \code{\link{varArrayTest}} for more details).
The log-contrast is specified by the \code{lc} argument using a vector of codes 1, -1 and 0 for components
in the numerator, denominator and omitted respectively.
}
\examples{
data(Water)
zPatterns(Water, label = 0)

# Test of homogeneity in log-contrast Potassium/Arsenic*Calcium
lcTest(Water, label = 0, lc = c(1,-1,-1,0))
}
\seealso{
\code{\link{zPatterns}}, \code{\link{varArray}}, \code{\link{varArrayError}}.
}

