% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrSVD.R
\name{lrSVD}
\alias{lrSVD}
\title{Log-ratio SVD algorithm}
\usage{
lrSVD(
  X,
  label = NULL,
  dl = NULL,
  frac = 0.65,
  ncp = 2,
  beta = 0.5,
  scale = FALSE,
  method = "Regularized",
  row.w = NULL,
  coeff.ridge = 1,
  threshold = 1e-04,
  seed = NULL,
  nb.init = 1,
  maxiter = 1000,
  ...
)
}
\arguments{
\item{X}{Compositional data set (\code{\link{matrix}} or \code{\link{data.frame}} class).}

\item{label}{Unique label (\code{\link{numeric}} or \code{\link{character}}) used to denote unobserved values in \code{X}.}

\item{dl}{Numeric vector or matrix of detection limits/thresholds. These must be given on the same scale as \code{X}.}

\item{frac}{Parameter for initial multiplicative simple replacement of left-censored data (see \code{\link{multRepl}}) (default = 0.65).}

\item{ncp}{PARAM_DESCRIPTION, Default: 2}

\item{beta}{PARAM_DESCRIPTION, Default: 0.5}

\item{scale}{PARAM_DESCRIPTION, Default: FALSE}

\item{method}{PARAM_DESCRIPTION, Default: 'Regularized'}

\item{row.w}{PARAM_DESCRIPTION, Default: NULL}

\item{coeff.ridge}{PARAM_DESCRIPTION, Default: 1}

\item{threshold}{PARAM_DESCRIPTION, Default: 1e-04}

\item{seed}{PARAM_DESCRIPTION, Default: NULL}

\item{nb.init}{PARAM_DESCRIPTION, Default: 1}

\item{maxiter}{PARAM_DESCRIPTION, Default: 1000}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
This function implements an algorithm to impute left-censored data (e.g. values below detection limit,
rounded zeros) based on the singular value decomposition (SVD) of a compositional data set.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
